feat(board): add basic types
The new types will be used for piece represensation, board state and
movement calculation. Following the ideas discussed in the Chess
Programming wiki, we are setting up to use bitboards with binary
oprations to have faster lookup.

We are defining the custom types `Piece`, `Color` and `Square` as
`uint8` and some custon constants to be used as enums (`iota`)

We also defined a smal basic struct for the Board that will use the
Bitboard type to represent state.