feat(board): add helper functions to Bitboard
The new functions will be used for baord manipulations.

`Set` is responsible for marking a square as ocupied.
`Clear` marks a square as free.
`IsSet` checks the current state of the board in a given square.
We also added a String representation method for debugging purposes.

test(board): add test cases for new Bitboard helper functions