feat(internal/board): add basics of move gen
Added representations for the offsets of different types of moves
(Sliding moves and knight and king moves).
The idea is that we will loop through the pieces and generate the
possible movements for each of then following the rules.

Missing the king and pawn movements still

TODO: Some extra boundary check for board wrapping.