fix(internal/board): stop edge wrap for sliding
Sliding moves, in particular diagonal ones would sometimes wrap around
the edge of the board. That would generate illegal moves and now the
checks have beem simplified for this purpose.

We also fixed the bishop test cases, as the generated move to `a8`
was valid but not listed as expected.