test(internal/board): add Pawn move test cases
BREAKING CHANGE: The current implementation of MoveList is a map that
does not allow duplicates, ad its is keyed by the target position of a
move. Therefore, Pawn promotion is not possible as it makes overrides
the contents for each possible promotion. A new implementation to fix
this is necessary

[skip ci]