fix(internal/board): tests for pawn moves
as pawn moves can involve promotion cases, that can result in a
situation where we have 4 moves all mapping to the same destination
square. With the previous approahc, this would result in an override of
the moves extracted to verify the behaviour of the code by the test.
With this enw approach, we use a composite key (destination, move type
and promotion piece) making ti possible to verify the pawn moves
generated properly.

as this is a special case for pawns, we are only using this custom
extraction inside the pawn test cases.