feat(internal/board): add PlayMove function
this function will be responsible for playing the move in the board. The
function takes a move and checks if the piece at the starting location
is valid (non empty, matching `SideToMove`). If so, the move will be
played else. Else, a custom error message is returned.

Currently only handling `Normal` move types.